email dynamic rework
