+++ packages/osc-api/emails/types.ts
@@ -1,0 +1,22 @@
+type regEmailData = {
+    to: string;
+    url: string;
+};
+
+type message = {
+    to: string;
+    from?: string;
+    sendId: string;
+    cc?: string[];
+    bcc?: string[];
+    replyTo?: string[];
+};
+
+type emailData = {
+    emailId: number;
+    message: message;
+    customProperties: {};
+};
+
+export type GetEmailData = (input: emailData) => Promise<String>;
+export type GetRegEmailData = (input: regEmailData) => Promise<String>;
