+++ packages/osc-api/index.ts
@@ -2,3 +2,4 @@ export * from './auth';
 export * from './user';
 export * from './course';
 export * from './utils';
+export * from './emails';
